Introduce pmap_change_prot().
Generalize pmap_change_attr_locked() so that it can be used to update
mappings permissions as well.  This is intended for use in the kernel
linker, both to apply protections to preloaded kernel modules (which are
not covered in kernel_map), and to ensure that the direct map is updated
when some protection changes are applied to the kernel map.