Add a KLD_SHARED variable to kmod.mk.
This can be used to override the platform default, since a few of our
drivers contain non-PIC blobs.  Set KLD_SHARED=no in the drivers which
require this.

Switch amd64 to KLD_SHARED by default.

Also set -fno-plt on platforms where KLDs are compiled as PIC.  This
reduces the overhead of calls to global functions in KLDs by eliminating
PLT stubs.