Formalize the use of linker scripts for kernel modules.
Automatically use ldscript.kmod.${MACHINE_CPUARCH} if it exists.
We already have an i386-specific linker script; rename it accordingly.
I plan to add an amd64 linker script to insert padding after .text
and .rodata*, so that permission boundaries are page-aligned.  (It turns
out that my plan to use DSOs on amd64 will not work because of issues
that arise when combining -mcmodel=kernel and -fPIC.)

Note that the linker script is applied during the partial link.  For
amd64 (which uses .o's), this is also the final link.  I think this is
probably fine, though maybe it will be useful someday to be able to
script the final link instead.