Add an ldscript for amd64 kmods.
Provide padding for .text* and .rodata* so that memory protections may
be enforced.

This is not perfect: some linker sets are read-only and live in the same
page as writeable linker sets.  For now I think it is reasonable to
simply live with this.

In the future it might be desireable to apply a heuristic to the .text
size of a KLD and instead pad to a 2MB boundary so that it gets mapped
with superpages.  For instance, we might do this if the size of .text is
greater than 1MB, as it is for zfs.ko.