Start implementing queue state updates using fcmpset loops. [4/5]
Convert the inactive queue scan.  Recall that the inactive queue scan,
unlike active and laundry queue scans, speculatively dequeues pages
before examining them.

As in the active queue scan, update act_count and potentially activate
or requeue the page if references were discovered.

Also modify vm_pageout_reinsert_inactive(), which requeues pages that
were found to be busy.  Simplify it so that it is not reimplementing
portions of vm_pqbatch_process_page().  In particular, if PGA_REQUEUE or
PGA_REQUEUE_HEAD is set, let that be handled during batch processing or
by the page daemon during a subsequent scan.