Optimize keg_drain().
Avoid holding the keg lock for longer than necessary.  This helps ensure
that trimming the zone doesn't block allocations for longer than really
needed.  If a zone's bucket list has recently been drained and we are
draining the keg, it is likely that allocations will block until
keg_drain() has emptied the free slab list, and that may take a long
time.