Rescue dumper support for the arm64 kernel.
This allows a kernel to save a minidump following a panic, by kexec'ing
an embedded "rescue" kernel, which boots with pre-reserved memory into
an embedded root image containing a script.  The script uses some
facilities of the rescue kernel to save a minidump from the panicked
kernel in the root filesystem, after which it reboots.

I tried to make this as self-contained as possible so as to make
upstreaming more straightforward.  Two new kernel options are added:
- RESCUE_SUPPORT: add the ability to exec a rescue kernel in order to
  obtain a minidump.  One must also set RESCUE_EMBED to the path of a
  compiled rescue kernel.
- RESCUE: a kernel which "rescues" a panicked kernel (compiled with
  RESCUE_SUPPORT) specifies this option.
In typical usage, GENERIC sets RESCUE_SUPPORT and the RESCUE_EMBED make
option, which is functionally the same as MFS_IMAGE.  RESCUE_EMBED
points to a kernel compiled with the RESCUE option.  I provided a sample
rescue kernel config; it just eliminates a few options from GENERIC to
save memory, but this is not strictly necessary in practice with a 64MB
reservation.

To enable rescue dumps, the GENERIC kernel must set the
"debug.rescue_minidump" tunable to "1".  If enabled, during boot the
GENERIC kernel will reserve 64MB of contiguous memory for use by the
rescue kernel.  The rescue kernel image is copied into the reserved
memory, along with some additional data needed to bootstrap.  No
modification to the DTB is required.  The kernel will automatically
create a dump device, so that the kernel will call into the minidump
code during a panic.  Once minidump parameters are calculated, it jumps
into the rescue "kexec" code, which finalizes bootstrap parameters,
sanitizies the execution environment and hands execution to the rescue
kernel that was copied into the memory reservation.

The implementation is somewhat coupled to the internals of locore.  It
also assumes that the embedded rescue kernel is an ELF file.  The rescue
kernel must be compiled without "options SMP", or
debug.rescue.kern.smp.disabled must be set in the host.  The former is
preferable to avoid wasting memory.

The rescue kernel boots like any other arm64 kernel, except that it must
synthesize some module metadata in initarm().  It provides a device
file, /dev/dumper, which can be read to obtain a minidump for the
panicked kernel.  The device is seekable and can be read multiple times
without modifying host memory.  The idea is to be able to simply read
the minidump using dd(1).

Tunables can be set in the rescue kernel by setting the same tunables in
the host with the prefix "debug.rescue.".  For example, to set
debug.debugger_on_panic="0" in the rescue kernel, set
debug.rescue.debug.debugger_on_panic="0" in the host.

I provided an rc script and minimal set of utilities which are
sufficient to recover a minidump.  The utilities are hardlinks of a
cross-compiled rescue(8) executable; the result fits in 8MB.  The rc
script looks for the rescue_rootfs tunable (specified as
debug.rescue.rescue_rootfs in the host), runs fsck_ffs on the specified
rootfs, mounts it, and saves the vmcore to /var/crash.