Remove the FIRMWARE_MAX limit.
The firmware module arbitrarily limits us to at most 50 images.  It is
possible to hit this limit on platforms that preload many firmware
images, or link all of the firmware images for a set of devices into the
kernel.

Convert the table into a linked list instead of blindly bumping the
maximum.