Fix UMA's first-touch policy with empty domains.
Suppose we are running on a CPU in a domain with no physical RAM.  When
an item is freed to a first-touch zone, it ends up in the cross-domain
bucket.  When the bucket is full, it gets placed in another domain's
bucket queue.  However, when allocating an item, UMA will always go to
the keg on a per-CPU cache miss because the empty domain's bucket queue
will always be empty.  This means that non-empty domains' bucket queues
can grow very rapidly on such systems.  For example, it can easily cause
mbuf allocation failures when the zone limit is reached.

Change cache_alloc() to follow a round-robin policy when running on an
empty domain.