uma: Respect uk_reserve in keg_drain().
When we set a reserve of free items for a zone, we must take care to
avoid reclamining them.  The two existing uma_zone_reserve() callers in
the tree fail to set UMA_ZONE_NOFREE, which would have the desired
effect.  Moreover, _NOFREE is a fairly large hammer.

Modify keg_drain() to simply respect the reserved pool.  For now we take
uk_reserve to be a per-NUMA domain quantity, since that's what existing
code does.

While here remove an always-false uk_freef == NULL check, and make sure
that the keg_drain() KTR statement does not reference an uninitialized
variable.